Feature/require auth

+++ src/routes/plants.js
@@ -12,6 +12,7 @@ import {
   getAvailablePlants,
   getAllPlants,
 } from "../db/plants.js";
+import { requireAuth } from "../middleware/auth.js";
 
 const plantRouter = Router();
 
@@ -25,7 +26,8 @@ plantRouter.get("/", async (req, res) => {
 });
 
 // GET /plants/all with search
-plantRouter.get("/all", async (req, res) => {
+//requireAuth
+plantRouter.get("/all", requireAuth, async (req, res) => {
   const { q } = req.query;
 
   const plants = await getAllPlants(q);
@@ -34,7 +36,7 @@ plantRouter.get("/all", async (req, res) => {
 });
 
 // GET /plants/:slug
-plantRouter.get("/:slug", async (req, res) => {
+plantRouter.get("/:slug", requireAuth, async (req, res) => {
   const slug = req.params.slug;
 
   const plant = await getPlantBySlug(slug);
@@ -44,25 +46,40 @@ plantRouter.get("/:slug", async (req, res) => {
       message: "Plant not found",
     });
   }
+
   res.json(plant)
 })
 
 // POST /plants
-plantRouter.post("/", validatePlant, validatePlantResult, async (req, res) => {
-  // TODO Validation for User and Admin
-  // validatePlant, validatePlantResult,
+plantRouter.post("/", requireAuth, validatePlant, validatePlantResult, async (req, res) => {
+  const plantData = {
+    ...req.body,
+    ownerId: req.userId 
+  };
 
-  const plant = await createPlant(req.body);
+  const plant = await createPlant(plantData);
 
   res.status(201).json(plant);
 });
 
 // PUT /plants/:slug
-plantRouter.put("/:slug", validatePlant, validatePlantResult, async (req, res) => {
-  // TODO Validation for User (owner) and Admin
-
+plantRouter.put("/:slug", requireAuth, validatePlant, validatePlantResult,
+  async (req, res) => {
+  // TODO Validation for Admin
   const slug = req.params.slug;
-  const { name, image, species, lightLevels, coordinates, meetingTime, ownerId } = req.body;
+
+  const plant = await getPlantBySlug(slug);
+
+   if(!plant){
+      return res.status(404).json({ message: "Plant not found" })
+  }
+
+  // Kontrollera att användaren äger plantan
+  if (plant.ownerId._id.toString() !== req.userId) {
+    return res.status(403).json({ message: "Not allowed to update this plant" })
+  }
+
+  const { name, image, species, lightLevels, coordinates, meetingTime } = req.body;
 
   const updatedPlant = await updatePlantBySlug(slug, {
     name,
@@ -83,9 +100,21 @@ plantRouter.put("/:slug", validatePlant, validatePlantResult, async (req, res) =
 });
 
 // PATCH /plants/:slug
-plantRouter.patch("/:slug", validatePlantUpdate, validatePlantResult, async (req, res) => {
+plantRouter.patch("/:slug", requireAuth, validatePlantUpdate, validatePlantResult, async (req, res) => {
   // TODO Validation for User (owner) and Admin
   const slug = req.params.slug
+
+  const plant = await getPlantBySlug(slug);
+
+   if(!plant){
+      return res.status(404).json({ message: "Plant not found" })
+  }
+
+  // Kontrollera att användaren äger plantan
+  if (plant.ownerId._id.toString() !== req.userId) {
+    return res.status(403).json({ message: "Not allowed to update this plant" })
+  }
+
   const { ownerId, slug: bodySlug, ...updateData } = req.body
 
   const updatedPlant = await updatePlantBySlug(slug, updateData)
@@ -100,13 +129,23 @@ plantRouter.patch("/:slug", validatePlantUpdate, validatePlantResult, async (req
 })
 
 // DELETE /plants/:slug
-plantRouter.delete("/:slug", async (req, res) => {
+plantRouter.delete("/:slug", requireAuth, async (req, res) => {
   // TODO Validation for User (owner) and Admin
 
   const slug = req.params.slug;
 
+  const findPlant = await getPlantBySlug(slug);
+  
+  if(!findPlant){
+    return res.status(404).json({ message: "Plant not found" })
+  }
+  
+  // Kontrollera att användaren äger plantan
+  if (findPlant.ownerId._id.toString() !== req.userId) {
+    return res.status(403).json({ message: "Not allowed to delete this plant" })
+  }
   const plant = await deletePlantBySlug(slug);
-
+  
   if (!plant) {
     return res.status(400).json({
       message: "Plant does not exist",
